requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
